requirements.txt
# Flask et extensions
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Marshmallow==0.15.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-RESTX==1.3.0

# Base de données
psycopg2-binary>=2.9.9

# Utilitaires
python-dotenv==1.0.0
Werkzeug==2.3.7

# Sécurité et validation
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Serveur de production
gunicorn==21.2.0

# Tests (optionnel)
pytest==7.4.2
pytest-flask==1.3.0
